requirements/base.txt
asgiref==3.10.0
django==5.2.7
pip==24.3.1
sqlparse==0.5.3
django-environ==0.12.0
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
pytz==2025.2
feedgen==1.0.0
qrcode==8.2
pillow==12.0.0